+++ resources/lang/en/lang.php
@@ -95,6 +95,9 @@ return [
         'nothing' => 'Nothing selected',
         'selected' => 'selected',
         'placeholder' => 'Select from the list',
+        'init'        => 'Press Enter to select',
+        'limit'       => 'and ${count} more',
+        'deselect'    => 'Press Enter to deselect',
     ],
     'image' => [
         'browse' => 'Select Image',