+++ resources/views/default/form/element/multiselect.blade.php
@@ -7,9 +7,31 @@
         @endif
     </label>
 
-    <div>
-        {!! Form::select($name, $options, $value, $attributes) !!}
-    </div>
+    <deselect :value="{{json_encode($value)}}" :multiple="true" :options="{{json_encode($options)}}" inline-template>
+        <div>
+            <multiselect v-model="val"
+                         track-by="id"
+                         label="text"
+                         :multiple="multiple"
+                         :limit="3"
+                         :searchable="true"
+                         :options="options"
+                         placeholder="{{ trans('sleeping_owl::lang.select.placeholder') }}"
+                         :select-label="'{{trans('sleeping_owl::lang.select.init')}}'"
+                         :selected-label="'{{trans('sleeping_owl::lang.select.selected')}}'"
+                         :deselect-label="'{{trans('sleeping_owl::lang.select.deselect')}}'"
+            >
+            </multiselect>
+
+            <select v-show="true == false" multiple name="{{$name}}">
+
+                <option :selected="hasOption(opt.id)" :value="opt.id"
+                        v-for="opt in options">
+                    @{{ opt.text }}
+                </option>
+            </select>
+        </div>
+    </deselect>
 
     @include(AdminTemplate::getViewPath('form.element.partials.helptext'))
     @include(AdminTemplate::getViewPath('form.element.partials.errors'))