+++ resources/views/default/form/element/select.blade.php
@@ -7,9 +7,24 @@
         @endif
     </label>
 
-    <div>
-        {!! Form::select($name, $options, $value, $attributes) !!}
-    </div>
+    <deselect :value="{{json_encode($value)}}" :multiple="false" :options="{{json_encode($options)}}" inline-template>
+        <div>
+            <multiselect v-model="val"
+                         track-by="id"
+                         label="text"
+                         :multiple="multiple"
+                         :searchable="true"
+                         :options="options"
+                         placeholder="{{ trans('sleeping_owl::lang.select.placeholder') }}"
+                         :select-label="'{{trans('sleeping_owl::lang.select.init')}}'"
+                         :selected-label="'{{trans('sleeping_owl::lang.select.selected')}}'"
+                         :deselect-label="'{{trans('sleeping_owl::lang.select.deselect')}}'"
+            >
+            </multiselect>
+
+            <input type="hidden" name="{{$name}}" v-model="preparedVal">
+        </div>
+    </deselect>
 
     @include(AdminTemplate::getViewPath('form.element.partials.helptext'))
     @include(AdminTemplate::getViewPath('form.element.partials.errors'))