+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 <p align="center"><h2>[Unreleased] (Only in <code class=" language-php">development</code> branch)</h2></p>
  
+ * [Feature] Изменились селекты. Теперь по умолчанию стоит [vue-multiselect](http://monterail.github.io/vue-multiselect) 
+             Очень не завидую тем у кого появлялись кастомные темплейты на элементы AdminFormElement::select и AdminFormElement::multiselect
+             их придется слегка подпилить. 
+             Оставляйте фидбеки.
  * [Feature] Появилась возможность редактировать JSON поля. Установка атрибутов идет в AtributeName через `field->key`
  * [Feature] Появилась пара новых эвентов
      * datatables::actions::submitting - эвент до действия после согласия SweetAlert в масс экшнс